requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
selenium==4.15.2
pillow==10.1.0